feat: add limit -1 for query (#3658)
The query `size` works like this, it should more clear now:
1. size > 0, it will return hits equal to `size`
2. size = 0, it will return empty hits, this used to other purposes,
like: get total, get aggs.
3. size = -1, it will use a backend default limit to return hits,
setting by `ZO_QUERY_DEFAULT_LIMIT=1000`

+++ src/handler/http/request/search/mod.rs
@@ -447,7 +447,7 @@ pub async fn around(
 
     let around_size = query
         .get("size")
-        .map_or(10, |v| v.parse::<usize>().unwrap_or(10));
+        .map_or(10, |v| v.parse::<i64>().unwrap_or(10));
 
     let regions = query.get("regions").map_or(vec![], |regions| {
         regions
@@ -934,7 +934,7 @@ async fn values_v1(
 
     let size = query
         .get("size")
-        .map_or(10, |v| v.parse::<usize>().unwrap_or(10));
+        .map_or(10, |v| v.parse::<i64>().unwrap_or(10));
     let start_time = query
         .get("start_time")
         .map_or(0, |v| v.parse::<i64>().unwrap_or(0));
@@ -1181,7 +1181,7 @@ async fn values_v2(
 
     let size = query
         .get("size")
-        .map_or(10, |v| v.parse::<usize>().unwrap_or(10));
+        .map_or(10, |v| v.parse::<i64>().unwrap_or(10));
     let start_time = query
         .get("start_time")
         .map_or(0, |v| v.parse::<i64>().unwrap_or(0));
